refactor sync file command

+++ pkg/content/content.go
@@ -17,15 +17,18 @@ import (
 )
 
 const (
-	JCRRoot              = "jcr_root"
-	JCRContentFile       = ".content.xml"
-	JCRContentFileSuffix = ".xml"
-	JCRMixinTypesProp    = "jcr:mixinTypes"
-	JCRRootPrefix        = "<jcr:root"
-	PropPattern          = "^\\s*([^ =]+)=\"([^\"]+)\"(.*)$"
-	NamespacePattern     = "^\\w+:(\\w+)=\"[^\"]+\"$"
-	ParentsBackupSuffix  = ".bak"
-	JCRContentNode       = "jcr:content"
+	JCRRoot                   = "jcr_root"
+	JCRContentFile            = ".content.xml"
+	JCRContentFileSuffix      = ".xml"
+	JCRMixinTypesProp         = "jcr:mixinTypes"
+	JCRRootPrefix             = "<jcr:root"
+	PropPattern               = "^\\s*([^ =]+)=\"([^\"]+)\"(.*)$"
+	NamespacePattern          = "^\\w+:(\\w+)=\"[^\"]+\"$"
+	ParentsBackupSuffix       = ".bak"
+	ParentsBackupDirIndicator = ".bakdir"
+	JCRContentNode            = "jcr:content"
+	JCRContentPrefix          = "<jcr:content"
+	JCRContentDirName         = "_jcr_content"
 )
 
 type Manager struct {
@@ -104,6 +107,9 @@ func (c Manager) CleanFile(path string) error {
 		return err
 	}
 	if err := c.flattenFile(path); err != nil {
+		return err
+	}
+	if err := deleteEmptyDirs(filepath.Dir(path)); err != nil {
 		return err
 	}
 	return nil
@@ -348,13 +354,25 @@ func (c Manager) doRootBackup(root string) error {
 
 func (c Manager) undoParentsBackup(root string) error {
 	return eachParentFiles(root, func(parent string) error {
+		indicator := false
+		if err := eachFilesInDir(parent, func(path string) error {
+			indicator = indicator || strings.HasSuffix(path, ParentsBackupDirIndicator)
+			return nil
+		}); err != nil {
+			return err
+		}
+		if !indicator {
+			return nil
+		}
+
 		if err := eachFilesInDir(parent, func(path string) error {
 			return deleteFile(path, func() bool {
-				return !strings.HasSuffix(path, ParentsBackupSuffix)
+				return !strings.HasSuffix(path, ParentsBackupSuffix) || !strings.HasSuffix(path, ParentsBackupDirIndicator)
 			})
 		}); err != nil {
 			return err
 		}
+
 		return eachFilesInDir(parent, func(path string) error {
 			if strings.HasSuffix(path, ParentsBackupSuffix) {
 				origin := strings.TrimSuffix(path, ParentsBackupSuffix)
@@ -433,6 +451,13 @@ func writeLines(path string, lines []string) error {
 }
 
 func (c Manager) backupFile(path string, format string) error {
+	dir := filepath.Dir(path)
+	indicator, err := os.Create(filepath.Join(dir, ParentsBackupDirIndicator))
+	if err != nil {
+		return err
+	}
+	defer func() { _ = indicator.Close() }()
+
 	source, err := os.Open(path)
 	if err != nil {
 		return err
@@ -478,4 +503,18 @@ func determineStringSlice(values any, key string) []string {
 		result = cast.ToStringSlice(value)
 	}
 	return result
+}
+
+func IsContentFile(path string) bool {
+	inputLines, err := readLines(path)
+	if err != nil {
+		return false
+	}
+
+	for _, inputLine := range inputLines {
+		if strings.Contains(inputLine, JCRContentPrefix) {
+			return true
+		}
+	}
+	return false
 }
